requirements.txt
alabaster==0.7.11
altair==4.1.0
Babel==2.6.0
bokeh==0.13.0
boto==2.49.0
google-api-core==1.26.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-datastore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googledrivedownloader==0.4
ipywidgets==7.4.1
matplotlib==3.3.2
nltk==3.3
openpyxl==2.5.6
pandas==0.23.4
pathlib2==2.3.2
plotly==4.14.3
PeakUtils==1.3.2
pyspellchecker==0.6.2
requests==2.25.1
requests-oauthlib==1.3.0
xlrd==1.1.0
xlsxwriter==1.1.0
scikit-image==0.14.0
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.9.0
sklearn==0.0
streamlit==0.80.0
tokenizers==0.8.1rc1
torch==1.8.1